Do not use sub classes for Broadcast Receivers
